backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.3.2
python-dotenv==1.0.0
supabase==2.3.4
pypdf==4.0.1
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
asyncpg==0.29.0
langchain-text-splitters==0.3.5
